5/8 homework uploaded
Check whether the email entered by the user is in the database. If there is, it will be displayed on the web page that it has been subscribed. If not, add this email to the database, and finally realize the function of uniformly sending electronic newspapers to all emails in the database.